back/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
Pillow==10.3.0
python-slugify==8.0.4
nanoid==2.0.0
